Create README.md
This project allows users to input their weight, height, and gender to calculate their Body 
Mass Index (BMI). Based on the BMI, the project determines whether the user needs to gain 
or lose weight and calculates their maintenance calories. 
For weight reduction, the project adjusts the calorie target to be less than the maintenance 
calories, and for weight gain, it adjusts the target to be more than the maintenance calories. 
Users specify how much weight they want to gain or lose. The project then provides a diet 
plan based on these goals. 
The knapsack algorithm is used to select the optimal foods within the adjusted calorie limits, 
ensuring that users receive the best nutritional value while staying within their calorie 
constraints. This approach allows users to achieve their dietary goals in a cost-effective way, 
getting a balanced diet that supports either weight loss or gain. 
The knapsack algorithm helps by selecƟng the best combinaƟon of foods that meet the targeted 
calorie intake, ensuring that the diet is both nutrious and effective.